backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9      # for form-data (file upload, login forms)
passlib[bcrypt]==1.7.4       # for password hashing
bcrypt==4.1.2                # required backend for passlib
sqlalchemy==2.0.34           # ORM for database
databases==0.9.0             # async db support
psycopg2-binary==2.9.9       # PostgreSQL driver (remove if using SQLite/other DB)
pydantic[email]==2.9.2       # validation (email, password, etc.)
python-jose[cryptography]
google-cloud-firestore
firebase_admin